move a util to server only utils file
