+++ src/lib/utils.ts
@@ -1,7 +1,6 @@
 import React from 'react';
 import { clsx, type ClassValue } from 'clsx';
 import { twMerge } from 'tailwind-merge';
-import { TRPCError } from '@trpc/server';
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -48,17 +47,4 @@ export function renderToText(node: React.ReactNode): string {
     return renderToText(el.props.children);
   }
   return '';
-}
-
-type IdRef = string | { id: string } | null | undefined;
-
-export function asId(ref: IdRef): string {
-  if (typeof ref === 'string') return ref;
-  if (ref && typeof ref === 'object' && typeof ref.id === 'string')
-    return ref.id;
-
-  throw new TRPCError({
-    code: 'BAD_REQUEST',
-    message: 'Missing or invalid tenant reference.'
-  });
 }
